Moved helper classes into a seperate directory
